% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COPtable.R
\name{COPtable}
\alias{COPtable}
\title{COPtable}
\usage{
COPtable(df, days, lagDays, measureTable, percentChangeKPI = 5)
}
\arguments{
\item{df}{Input data from \code{rollSummaryXDays()}. Either \code{cdtHosp7DayRollingData} or \code{cdtHosp14DayRollingData}}

\item{days}{The number of days that the time period is over. Should match the days argument for the \code{numDays} argument of \code{rollSummaryXDays()}.}

\item{lagDays}{The number of days to lag the data for.}

\item{measureTable}{A formatted data.frame contianing the columns 'measureName', 'Avg_Total', 'measureDisplayName', 'upGood', 'PerCapita'}

\item{percentChangeKPI}{What percent change should the KPI's be activated for? Default is 5\%.}
}
\description{
Create the dataset used by the COPtable with the multiple slicers
}
