% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBiDatasets.R
\name{createBiDatasets}
\alias{createBiDatasets}
\title{Create the Dataset for Power Bi}
\usage{
createBiDatasets(
  baseDataList = downloadAllCovidAPIData(),
  lagDaysCDT = 10,
  lagDaysHosp = 2
)
}
\arguments{
\item{baseDataList}{A named list of data.frames containing the base data. See
details for more information. Defaults to the return from
\code{downloadAllCovidAPIData()}}

\item{lagDaysCDT}{Number of days to lag the Case, Death, Test data. Defaults
to the value used by the Hub (10).}

\item{lagDaysHosp}{Number of days to lag the Hospital data. Defaults to the
value used by the Hub (2).}
}
\value{
A list of data.frames that are used by MARC's COVID Data Hub.
}
\description{
Given the base datasets, this function will do all the
  transformations and summarizations that are used to feed the data being
  displayed on the MARC's COVID Data Hub.
}
\details{
\code{baseDataList} should contain a named list of the base data.frames. These
  are available through the MARC data API through the helpful functions
  \code{downloadMARCCovidData()} and \code{downloadAllCovidAPIData()} This should be 3
  data.frames with he following names:
\describe{
  \item{cdtData}{Case, Death, and Test Data}
  \item{cdtNRData}{Newly Reported Case, Death, and Test Data}
  \item{hospData}{Hospital Data}
}
}
