% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadAllCovidAPIData}
\alias{downloadAllCovidAPIData}
\title{Download All of MARC's COVID-19 Data}
\usage{
downloadAllCovidAPIData()
}
\value{
A list of tibbles containing the COVID-19 data
}
\description{
Allows all the COVID-19 data to be downloaded from the MARC Data
  API with a simple function. This is a helper function for the data pipeline
  so that for testing purposes, data can be ingested directly from MARC
  internal SQL servers and not hard coded to the API data.
}
\examples{
\dontrun{

downloadAllCovidAPIData()

}
}
