% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadBaseData}
\alias{downloadBaseData}
\title{Download MARC's COVID-19 Data}
\usage{
downloadBaseData(
  type = c("CDT", "CDT_NewlyReported", "Hospital"),
  source = "MARCDataAPI",
  date = NULL
)
}
\arguments{
\item{type}{What table should be downloaded? Select one of 'CDT', 'CDT_NewlyReported', or 'Hospital'.}

\item{source}{What should the source of the data be? Defaults to MARCDataAPI. Other options are passed to `Covid19MARCInternal::downloadBaseDataInternal()` for internal use.}

\item{date}{Default NULL. Only used by `Covid19MARCInternal::downloadBaseDataInternal()` for internal use.}
}
\value{
A tibble with the selected data
}
\description{
Allows data to be downloaded from the MARC Data API with a
  simple function.
}
\details{
On the MARC internal network, the suggested Covid19MARCInternal
  package can be used to access the same data as provided through the API
  directly from the Publication and Production servers for testing purposes
}
\examples{
\dontrun{

cdtData <- downloadBaseData(type = "CDT")

cdtNRData <-  downloadBaseData(type = "CDT_NewlyReported")


}
}
