% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{downloadMARCCovidData}
\alias{downloadMARCCovidData}
\title{Download MARC's COVID-19 Data}
\usage{
downloadMARCCovidData(dataset = c("CDT", "CDT_NewlyReported", "Hospital"))
}
\arguments{
\item{dataset}{What dataset should be downloaded? Select one of 'CDT', 'CDT_NewlyReported', or 'Hospital'.}
}
\value{
A tibble with the selected COVID-19 data
}
\description{
Allows data to be downloaded from the MARC Data API with a
  simple function.
}
\examples{
\dontrun{

cdtData <- downloadBaseData(type = "CDT")

cdtNRData <-  downloadBaseData(type = "CDT_NewlyReported")


}
}
