% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadData.R
\name{getBaseCovidData}
\alias{getBaseCovidData}
\title{Create Base Calculated Datasets from Raw Data}
\usage{
getBaseCovidData(baseDataList = downloadAllCovidAPIData())
}
\arguments{
\item{baseDataList}{A named list of data.frames containing the base data. See
details for more information. Defaults to the return from
\code{downloadAllCovidAPIData()}}
}
\value{
A list of data.frames that are used by MARC's COVID Data Hub.
}
\description{
Create the Base Datasets that all of the COVID-19 Data Hub is ran from
}
\details{
\code{baseDataList} should contain a named list of the base data.frames. These
  are available through the MARC data API through the helpful functions
  \code{downloadMARCCovidData()} and \code{downloadAllCovidAPIData()} This should be 3
  data.frames with the following names:
\describe{
  \item{cdtData}{Case, Death, and Test Data}
  \item{cdtNRData}{Newly Reported Case, Death, and Test Data}
  \item{hospData}{Hospital Data}
}
}
