% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jurisTable}
\alias{jurisTable}
\alias{popTable}
\title{COVID-19 Jurisdiction Information}
\format{
A data frame with 30 rows and 7 variables:
\describe{
  \item{Jurisdiction}{Name of the county, city or region; character; [NVARCHAR](30)}
  \item{Region}{Name of the region that the jurisdiciton is apart of; character: one of 'HCC', 'MARC', 'State', 'StateMARC', 'MARC_HCC', 'HCCRegion'; [NVARCHAR](10)}
  \item{State}{Name of the state that the jurisdiciton is apart of; character: one of 'KS', 'MO', NA; [NVARCHAR](2)}
  \item{GeoID}{Unique GeoID code for the jurisdiciton (unique identifier); character; [NVARCHAR](10)}
  \item{Population}{Population in the jurisdiciton (source: ACS data); integer; [INT]}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 34 rows and 2 columns.
}
\source{
'Covid19MARCInternal::jurisdictionTable'
}
\usage{
jurisTable

popTable
}
\description{
A dataset containing keys for jurisdiction information used on the COVID-19 project.
}
\details{
\code{jurisTable} and \code{popTable} are just a subsets of \code{Covid19MARCInternal::jurisdictionTable} with specific columns filtered
}
\keyword{datasets}
