% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list2CSV.R
\name{list2CSV}
\alias{list2CSV}
\title{Export List to CSVs}
\usage{
list2CSV(dfList, outputFolder)
}
\arguments{
\item{dfList}{A named list of data.frames for export}

\item{outputFolder}{The file path to the folder that the CSV's will be exported to.}
}
\description{
Export List to CSVs
}
