% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mostRecentGivenTime.R
\name{mostRecentGivenTime_Vacc}
\alias{mostRecentGivenTime_Vacc}
\title{Most Recent Vaccine Data Given A Time Period}
\usage{
mostRecentGivenTime_Vacc(df, days, lagDays)
}
\arguments{
\item{df}{Input data.frame. It is currently absorbing 'bi_vaccMO_DailyData'.}

\item{days}{Number of days to use for the time frame. A value of NA will trigger the 'All Time' summary.}

\item{lagdays}{Number of days to lag the time frame.}
}
\description{
Used to create the jurisdiction bar charts with time and per
  capita/raw data slicers for the Vaccination power bi pages
}
\details{
Currently only set up to create summaries for the vaccination data in the time period.
}
