% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schoolGatingCriteria.R
\name{SGC_14DaySummary}
\alias{SGC_14DaySummary}
\title{Create School Gating Criteria Table}
\usage{
SGC_14DaySummary(df_14DayRolling, lagDays = 7)
}
\arguments{
\item{df_14DayRolling}{The 14 day output from `rollSummaryXDays()`.
'cdtHosp14DayRollingData' from `getBaseCovidData()`}

\item{lagDays}{The number of days to lag by. Defaults at 7 days which has
been the historic number used for the school gating criteria page.}
}
\description{
Given the 14 day rolling averages, format the columns needed to
  run the school gating criteria page.
}
