% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createBiDatasets_WDS.R
\name{createBiDatasets_WDS}
\alias{createBiDatasets_WDS}
\title{Create Weekly Data Snapshot Data for Power Bi}
\usage{
createBiDatasets_WDS(
  baseDataList = getBaseCovidData(),
  cutoffDay = "Sunday",
  lagDaysCDT = 10,
  lagDaysHosp = 2
)
}
\description{
Given the base datasets, this function will do all the
  transformations and summarizations that are used to feed the data being
  displayed on the Weekly Data Snapshot
}
