% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jurisdictionTable}
\alias{jurisdictionTable}
\alias{jurisdictionTableCDT}
\title{COVID-19 Jurisdiction Information}
\format{
A data frame with 30 rows and 7 variables:
\describe{
  \item{Jurisdiction}{Name of the county, city or region; character; [NVARCHAR](30)}
  \item{Region}{Name of the region that the jurisdiciton is apart of; character: one of 'HCC', 'MARC', 'State', 'StateMARC', 'MARC_HCC', 'HCCRegion'; [NVARCHAR](10)}
  \item{State}{Name of the state that the jurisdiciton is apart of; character: one of 'KS', 'MO', NA; [NVARCHAR](2)}
  \item{GeoID}{Unique GeoID code for the jurisdiciton (unique identifier); character; [NVARCHAR](10)}
  \item{Population}{Population in the jurisdiciton (source: ACS data); integer; [INT]}
  \item{CDT_Jurisdiction}{TRUE/FALSE on if this Jurisdiciton is used to enter Cases, Deaths, Tests (CDT) data; logical; [BIT]}
  \item{DisplayName}{Pretty jurisdiction name used for display purposes; character; [NVARCHAR](30)}
}

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 20 rows and 7 columns.
}
\source{
'/data-raw/CovidJurisdictions.csv'
}
\usage{
jurisdictionTable

jurisdictionTableCDT
}
\description{
A dataset containing keys for jurisdiction information used on the COVID-19 project.
}
\details{
\code{jurisdictionTableCDT} is just a subset of \code{jurisdictionTable} filtered to where 'CDT_Jurisdiction' is TRUE
}
\keyword{datasets}
